extract/dedup listFilesAtPath
