+++ pkg/csplugin/utils_windows_test.go
@@ -1,4 +1,4 @@
-/////go:build windows
+//go:build windows
 
 package csplugin
